+++ Android/UdacityExercises/Lesson09b-ToDo-List-AAC/T09b.09-Exercise-AddTheViewModel/app/src/main/java/com/example/android/todolist/MainViewModel.java
@@ -1,11 +1,30 @@
 package com.example.android.todolist;
 
-// TODO (1) make this class extend AndroidViewModel and implement its default constructor
-public class MainViewModel {
+import android.app.Application;
+import android.arch.lifecycle.AndroidViewModel;
+import android.arch.lifecycle.LiveData;
+import android.support.annotation.NonNull;
+import android.util.Log;
 
-    // TODO (2) Add a tasks member variable for a list of TaskEntry objects wrapped in a LiveData
+import com.example.android.todolist.database.AppDatabase;
+import com.example.android.todolist.database.TaskEntry;
 
-    // TODO (4) In the constructor use the loadAllTasks of the taskDao to initialize the tasks variable
+import java.util.List;
 
-    // TODO (3) Create a getter for the tasks variable
+public class MainViewModel extends AndroidViewModel {
+
+    private static final String TAG = MainViewModel.class.getSimpleName();
+
+    private LiveData<List<TaskEntry>> tasks;
+
+    public MainViewModel(@NonNull Application application) {
+        super(application);
+        AppDatabase database = AppDatabase.getInstance(this.getApplication());
+        Log.d(TAG, "Actively retrieving the tasks from the DataBase");
+        tasks = database.taskDao().loadAllTasks();
+    }
+
+    public LiveData<List<TaskEntry>> getTasks() {
+        return tasks;
+    }
 }